fetch/api/headers/headers-basic.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Headers structure</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#headers">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      var headerDict = {"name1": "value1",
                        "name2": "value2",
                        "name3": "value3"
      };

      var headerSeq = [];
      for (var name in headerDict)
        headerSeq.push([name, headerDict[name]]);

      test(function() {
        var headers = new Headers(headerSeq);
        for (name in headerDict)
          assert_equals(headers.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
      }, "Create headers with sequence");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_equals(headers.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
      }, "Create headers with OpenEndedDictionary");

      test(function() {
        var headers = new Headers(headerDict);
        var headers2 = new Headers(headers);
        for (name in headerDict)
          assert_equals(headers2.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
      }, "Create headers whith existing headers");

      test(function() {
        var headers = new Headers();
        for (name in headerDict)
          headers.append(name, headerDict[name]);
          assert_equals(headers.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
      }, "Check append method");

      test(function() {
        var headers = new Headers();
        for (name in headerDict) {
          headers.set(name, headerDict[name]);
          assert_equals(headers.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
        }
      }, "Check set method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_true(headers.has(name),"headers has name " + name);

        assert_false(headers.has("nameNotInHeaders"),"headers has not name: nameNotInHeaders");
      }, "Check has method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict) {
          assert_true(headers.has(name),"headers has name " + name);
          headers.delete(name)
          assert_true(!headers.has(name),"headers has not name " + name +" anymore");
        }
      }, "Check delete method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_equals(headers.get(name), headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);

        assert_equals(headers.get("nameNotInHeaders"), null, "name: nameNotInHeaders has no value");
      }, "Check get method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_equals(headers.getAll(name)[0], headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);

        for (name in headerDict) {
          headers.append(name, headerDict[name]+"-bis");
          assert_equals(headers.getAll(name)[0], headerDict[name],
            "name: " + name + "has value: " + headerDict[name]);
          assert_equals(headers.getAll(name)[1], headerDict[name]+"-bis",
            "name: " + name + "has also value: " + headerDict[name]+"-bis");
        }
        assert_equals(headers.getAll("nameNotInHeaders").length, 0 , "name: nameNotInHeaders has no value");
      }, "Check getAll method");
    </script>
  </body>
</html>
